+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.0.3 (unreleased)
 
+* Add `OPAAuthorizationManager` constructor that accepts a path but not an `OPAClient`.
+
 ## v0.0.2
 
 * Rather than hard-coding `en`, the preferred key to search for decision reasons for can now be changed with `OPAAuthorizationManager.setReasonKey()`. The default remains `en`.